Move main go to sample file.
